Add 'skipped' to download items and complete filter output if skipped
